Name-api/requirements.txt
autopep8==1.6.0
fastapi==0.88.0
httptools==0.5.0
pip-chill==1.0.1
python-dotenv==0.21.0
pyyaml==6.0
tensorflow-cpu==2.11.0
uvicorn==0.20.0
uvloop==0.17.0
watchgod==0.7
websockets==10.4
